+++ Crossing_Game_OOP/cWidget.cpp
@@ -68,7 +68,7 @@ void cDWindow::unshow(bool showNow)
 }
 
 
-cButton::cButton(cDWindow* parent, COORD offsetFromParentTopleft, const string& tagName, const string& imgSrc, int borderDensity, void (*pFunction) (void)) : cWidget(static_cast<cWidget*> (parent), offsetFromParentTopleft, tagName, imgSrc)
+cButton::cButton(cDWindow* parent, COORD offsetFromParentTopleft, const string& tagName, const string& imgSrc, short borderDensity, void (*pFunction) (void)) : cWidget(static_cast<cWidget*> (parent), offsetFromParentTopleft, tagName, imgSrc)
 {
 	bordDensity = borderDensity;
